shop_app/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coverage==7.3.0
crispy-bootstrap4==2022.1
cryptography==41.0.3
da-vinci==0.4.0
defusedxml==0.7.1
Django==4.2.3
django-admin-thumbnails==0.2.8
django-allauth==0.55.0
django-crispy-forms==2.0
django-ipware==5.0.0
django-recaptcha==3.0.0
django-session-timeout==0.1.0
django-tinymce==3.6.1
exceptiongroup==1.1.3
flake8==6.1.0
flake8-isort==6.0.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.1
pytest-cov==4.1.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
secure-smtplib==0.1.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
